30_AWS_SageMaker/Bird Classification/requirements.txt
torch==1.8.0
torchvision==0.9.0
pandas==1.2.4
numpy==1.20.2
Pillow==8.2.0
boto3==1.17.46
sagemaker==2.31.0
pytest==6.2.3
flake8==3.9.0